third_party/blink/web_tests/plugins/user-gesture-expected.txt

Plugin received event: MouseEnter
* not handling user gesture
Plugin received event: MouseMove
* not handling user gesture
Plugin received event: MouseDown
* handling user gesture
Plugin received event: MouseUp
* handling user gesture
Plugin received event: MouseUp
* handling user gesture
Test that the user gesture indicator is correctly set for events forwarded through the WebPluginContainer. The tests succeeds if the mousedown/mouseup event pair is a user gesture.
